requirements.txt
# INTENTIONALLY VULNERABLE - FOR SECURITY SCANNER TESTING ONLY
# Known CVEs to trigger Trivy / dependency scanners.
Django==3.1
Flask==1.0
requests==2.25.0
Jinja2==2.11.3
cryptography==3.4
Pillow==8.0
urllib3==1.26.4
paramiko==2.7.2
PyYAML==5.3.1
lxml==4.6.2
numpy==1.19.4
setuptools==49.6.0
certifi==2020.12.5
